+++ apps/gateway/src/chat/chat.module.ts
@@ -1,5 +1,6 @@
 import { Module } from "@nestjs/common";
-import { ClientsModule, Transport } from "@nestjs/microservices";
+import { ResilientClientTCP } from "@app/common";
+import { ClientsModule } from "@nestjs/microservices";
 import {
   NAME_SERVICE_TCP,
   PORT_TCP,
@@ -14,7 +15,7 @@ import { ChatWsGateway } from "./chat.ws-gateway";
     ClientsModule.register([
       {
         name: NAME_SERVICE_TCP.CHAT_SERVICE,
-        transport: Transport.TCP,
+        customClass: ResilientClientTCP,
         options: {
           host: TCP_HOST,
           port: PORT_TCP.CHAT_SERVICE_PORT,
@@ -22,7 +23,7 @@ import { ChatWsGateway } from "./chat.ws-gateway";
       },
       {
         name: NAME_SERVICE_TCP.USER_SERVICE,
-        transport: Transport.TCP,
+        customClass: ResilientClientTCP,
         options: {
           host: TCP_HOST,
           port: PORT_TCP.USER_TCP_PORT,