+++ apps/gateway/src/notification/notification.module.ts
@@ -1,5 +1,6 @@
 import { Module } from "@nestjs/common";
-import { ClientsModule, Transport } from "@nestjs/microservices";
+import { ResilientClientTCP } from "@app/common";
+import { ClientsModule } from "@nestjs/microservices";
 import {
   NAME_SERVICE_TCP,
   PORT_TCP,
@@ -15,7 +16,7 @@ import { NotificationPushController } from "./notification.push.controller";
     ClientsModule.register([
       {
         name: NAME_SERVICE_TCP.NOTIFICATION_SERVICE,
-        transport: Transport.TCP,
+        customClass: ResilientClientTCP,
         options: {
           host: TCP_HOST,
           port: PORT_TCP.NOTIFICATION_TCP_PORT,
@@ -23,12 +24,12 @@ import { NotificationPushController } from "./notification.push.controller";
       },
       {
         name: NAME_SERVICE_TCP.SOCIAL_SERVICE,
-        transport: Transport.TCP,
+        customClass: ResilientClientTCP,
         options: { host: TCP_HOST, port: PORT_TCP.SOCIAL_TCP_PORT },
       },
       {
         name: NAME_SERVICE_TCP.USER_SERVICE,
-        transport: Transport.TCP,
+        customClass: ResilientClientTCP,
         options: { host: TCP_HOST, port: PORT_TCP.USER_TCP_PORT },
       },
     ]),