+++ apps/gateway/src/payment-options/payment-options.module.ts
@@ -1,5 +1,6 @@
 import { Module } from "@nestjs/common";
-import { ClientsModule, Transport } from "@nestjs/microservices";
+import { ResilientClientTCP } from "@app/common";
+import { ClientsModule } from "@nestjs/microservices";
 import {
   NAME_SERVICE_TCP,
   PORT_TCP,
@@ -13,7 +14,7 @@ import { PaymentOptionsService } from "./payment-options.service";
     ClientsModule.register([
       {
         name: NAME_SERVICE_TCP.PAYMENT_SERVICE,
-        transport: Transport.TCP,
+        customClass: ResilientClientTCP,
         options: {
           host: TCP_HOST,
           port: PORT_TCP.PAYMENT_TCP_PORT,